requirements.txt
# Pip Requirements File
shodan 
confluent_kafka
python-dotenv